requirements.txt
# Core dependencies
anthropic>=0.30.0

# Local embeddings (completely free)
sentence-transformers>=2.2.0

# LangGraph for agent orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0

# Vector database
chromadb>=0.4.0

# Document processing
PyMuPDF>=1.23.0
markdown>=3.5
python-magic>=0.4.27
tiktoken>=0.5.0

# Async and utilities
aiofiles>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
